+++ packages/cli/commons/src/utils/ux.ts
@@ -1,9 +1,9 @@
 import {CliUx} from '@oclif/core';
+import isCi from 'is-ci';
 import {red, green, magenta} from 'chalk';
 
-
 function isWindows() {
-  return process.platform === 'win32'
+  return process.platform === 'win32';
 }
 
 function isWindowsTerminal() {
@@ -33,3 +33,6 @@ export function stopSpinner(options?: {success?: boolean}) {
 
 export const formatOrgId = (orgId: TemplateStringsArray | string) =>
   shouldUseColor() ? magenta(orgId) : orgId;
+
+export const confirm = (message: string, ciDefault: boolean) =>
+  isCi ? Promise.resolve(ciDefault) : CliUx.ux.confirm(message);