+++ packages/cli/core/src/commands/atomic/component.ts
@@ -2,10 +2,12 @@ import {CLICommand} from '@coveo/cli-commons/command/cliCommand';
 import {UnknownError} from '@coveo/cli-commons/errors/unknownError';
 import {Flags} from '@oclif/core';
 import inquirer from 'inquirer';
-import {appendCmdIfWindows} from '../../lib/utils/os';
-import {spawnProcess} from '../../lib/utils/process';
-import {startSpinner} from '@coveo/cli-commons/utils/ux';
+import npf from '@coveo/cli-commons/npm/npf';
 import {Trackable} from '@coveo/cli-commons/preconditions/trackable';
+import {getPackageVersion} from '../../lib/utils/misc';
+import {handleForkedProcess} from '../../lib/utils/process';
+import {SubprocessError} from '../../lib/errors/subprocessError';
+import {isAggregatedErrorLike} from '../../lib/utils/errorSchemas';
 
 export default class AtomicInit extends CLICommand {
   public static description =
@@ -31,18 +33,35 @@ export default class AtomicInit extends CLICommand {
   @Trackable()
   public async run(): Promise<void> {
     const {initializer, name} = await this.getSpawnOptions();
-
-    const cliArgs = ['init', initializer, name];
-    startSpinner('Scaffolding project');
-    await spawnProcess(appendCmdIfWindows`npm`, cliArgs);
+    const forkedProcess = npf(initializer, [name], {stdio: 'inherit'});
+    try {
+      await handleForkedProcess(forkedProcess);
+    } catch (error) {
+      if (isAggregatedErrorLike(error)) {
+        const stderrMessageParts = [error.message];
+        stderrMessageParts.push(
+          ...error.errors.map((suberror) => ` • ${suberror}`)
+        );
+        throw new SubprocessError(error.message, stderrMessageParts.join('\n'));
+      } else {
+        throw error;
+      }
+    }
   }
 
   private async getSpawnOptions() {
     const {args, flags} = await this.parse(AtomicInit);
     const type = flags.type || (await this.askType());
 
-    const initializer = this.getInitializerPackage(type);
-    return {initializer, name: args.name};
+    const initializerPackage = this.getInitializerPackage(type);
+
+    // TODO CDX-1340: Refactor the replace into a well named utils.
+    return {
+      initializer: `${initializerPackage}@${getPackageVersion(
+        initializerPackage
+      )}`,
+      name: args.name,
+    };
   }
 
   private async askType(): Promise<string> {
@@ -60,9 +79,9 @@ export default class AtomicInit extends CLICommand {
   private getInitializerPackage(type: string): string {
     switch (type) {
       case 'page':
-        return '@coveo/atomic-component';
+        return '@coveo/create-atomic-component';
       case 'result':
-        return '@coveo/atomic-result-component';
+        return '@coveo/create-atomic-result-component';
       default:
         throw new UnknownError();
     }