+++ packages/cli/core/src/commands/ui/create/react.ts
@@ -5,7 +5,7 @@ import {platformUrl} from '@coveo/cli-commons/platform/environment';
 import {AuthenticatedClient} from '@coveo/cli-commons/platform/authenticatedClient';
 import {spawnProcess, spawnProcessOutput} from '../../../lib/utils/process';
 import {getPackageVersion} from '../../../lib/utils/misc';
-import {appendCmdIfWindows} from '../../../lib/utils/os';
+import {appendCmdIfWindows} from '@coveo/cli-commons/utils/os';
 import {
   Preconditions,
   IsAuthenticated,
@@ -27,6 +27,7 @@ type ReactProcessEnv = {
   apiKey: string;
   user: string;
   platformUrl: string;
+  platformEnvironment: string;
 };
 
 export default class React extends CLICommand {
@@ -91,6 +92,7 @@ export default class React extends CLICommand {
       apiKey: apiKey.value!,
       user: username,
       platformUrl: platformUrl({environment: cfg.environment}),
+      platformEnvironment: cfg.environment,
     };
 
     await spawnProcessOutput(appendCmdIfWindows`npm`, ['run', 'setup-env'], {