+++ CHANGELOG.md
@@ -1,3 +1,49 @@
+# release-16 (2023-06-13)
+
+### Bug Fixes
+
+- **accd:** updates deps ([#1302](https://github.com/coveo/cli/issues/1302)) ([bebb09d](https://github.com/coveo/cli/commits/bebb09dca354cb33efaceffe844706d191a400c5))
+- **cli:** use getPackageVersion for create-atomic-component-project ([#1293](https://github.com/coveo/cli/issues/1293)) ([b045bfe](https://github.com/coveo/cli/commits/b045bfe713e1d7de9cc921703019c339778bdf2b)), closes [#1292](https://github.com/coveo/cli/issues/1292) [#1292](https://github.com/coveo/cli/issues/1292)
+- remove unnecessary `cd` instruction from success message ([#1303](https://github.com/coveo/cli/issues/1303)) ([f23c4eb](https://github.com/coveo/cli/commits/f23c4eb2d2b3fdf0c77e22f078396b76eadd32d1))
+- remove unnecessary privilege ([#1301](https://github.com/coveo/cli/issues/1301)) ([db73b3b](https://github.com/coveo/cli/commits/db73b3bf0e569d7bd7d6063f1339f7214c64c90d))
+
+### Features
+
+- **accd:** handle errors from initializers in the CLI ([#1316](https://github.com/coveo/cli/issues/1316)) ([aee4429](https://github.com/coveo/cli/commits/aee4429bb83362a426ddb3f1f1f7b81c6b3be367))
+- **commons:** add npf (Node Package Fork) ([#1312](https://github.com/coveo/cli/issues/1312)) ([9955ddd](https://github.com/coveo/cli/commits/9955ddde00b4c65db92ff9a0729e8039f7d9a902))
+- **oauth:** swap default and fallback port ([#1305](https://github.com/coveo/cli/issues/1305)) ([3bf1e7a](https://github.com/coveo/cli/commits/3bf1e7a08872e3953d75450ba6e6edca7e713ad3))
+- **organizationendpoints:** add support for organization endpoints for UI commands and projects ([#1289](https://github.com/coveo/cli/issues/1289)) ([3b7653d](https://github.com/coveo/cli/commits/3b7653dbf1b59015afb4575bd265ec0a91b2bcef))
+
+# release-15 (2023-04-21)
+
+### Bug Fixes
+
+- **atomic:** don't need peerDeps ([#1294](https://github.com/coveo/cli/issues/1294)) ([8a7e7a9](https://github.com/coveo/cli/commits/8a7e7a9ae60b3837d2e3820d1af824756e4db549))
+
+# release-14 (2023-04-21)
+
+### Bug Fixes
+
+- **atomic:** remove create- prefix when using npm init ([#1290](https://github.com/coveo/cli/issues/1290)) ([f32c9fa](https://github.com/coveo/cli/commits/f32c9fa2d359fb206105a8d4ea144e772a68d763))
+- **atomic:** remove spinner in `atomic:cmp` ([#1291](https://github.com/coveo/cli/issues/1291)) ([68b4e06](https://github.com/coveo/cli/commits/68b4e0666924730e66986a8ddab4497dc27be3fb))
+- **atomic:** use the initializer version from CLI ([#1292](https://github.com/coveo/cli/issues/1292)) ([339925b](https://github.com/coveo/cli/commits/339925b5f187a7673e4eab649c50f952497c0c71))
+- **atomic:** wait for atomic to load/init on page component ([#1284](https://github.com/coveo/cli/issues/1284)) ([0113bf6](https://github.com/coveo/cli/commits/0113bf6e0b11cd47e6af988afc1434ea9f95cfbc))
+- **cli-core,cli-source,ui-react:** enforce LTS ([#1275](https://github.com/coveo/cli/issues/1275)) ([46cfeb2](https://github.com/coveo/cli/commits/46cfeb24c70ea96467e1d705537790dfaa929957))
+- **cli:** do not reassign fetch after polyfill ([#1271](https://github.com/coveo/cli/issues/1271)) ([85725f8](https://github.com/coveo/cli/commits/85725f8310e3395450b77c1a55939d7d9abfa3fe))
+- **cli:** use proper package ([#1280](https://github.com/coveo/cli/issues/1280)) ([9324725](https://github.com/coveo/cli/commits/93247255d6d7d38265c7e62b7e3156f266fa7552))
+- **hostedpages:** add explanation about changes done in the header of the local html pages ([#1263](https://github.com/coveo/cli/issues/1263)) ([5f2bfa4](https://github.com/coveo/cli/commits/5f2bfa49643f70461723c5dbd6d9c7f01fe68032))
+
+### Features
+
+- add confirmation message upon component creation ([#1270](https://github.com/coveo/cli/issues/1270)) ([e7294ce](https://github.com/coveo/cli/commits/e7294ce591b4c5b62ae63c3037b4ceb8c69f2d87))
+- **atomic:** release of ACCD ([#1287](https://github.com/coveo/cli/issues/1287)) ([5ce3cda](https://github.com/coveo/cli/commits/5ce3cda28a7a68ec1cef34e49f60b8f1e82da4b0))
+- **atomic:** remove beta tags ([#1283](https://github.com/coveo/cli/issues/1283)) ([4df0d2f](https://github.com/coveo/cli/commits/4df0d2fba2a182df78181e7611b880cf772358a9))
+- sanitize custom component name on creation ([#1272](https://github.com/coveo/cli/issues/1272)) ([1c434b6](https://github.com/coveo/cli/commits/1c434b6c83fe688f37ff843c90aa01e84a5a1129))
+
+### BREAKING CHANGES
+
+- **atomic:** ACCD is now stable.
+
 # release-13 (2023-04-05)
 
 ### Bug Fixes