+++ packages/cli/source/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.11 (2023-06-13)
+
+## 2.0.10 (2023-04-21)
+
+### Bug Fixes
+
+- **atomic:** don't need peerDeps ([#1294](https://github.com/coveo/cli/issues/1294)) ([8a7e7a9](https://github.com/coveo/cli/commits/8a7e7a9ae60b3837d2e3820d1af824756e4db549))
+
+## 2.0.9 (2023-04-21)
+
+### Bug Fixes
+
+- **cli-core,cli-source,ui-react:** enforce LTS ([#1275](https://github.com/coveo/cli/issues/1275)) ([46cfeb2](https://github.com/coveo/cli/commits/46cfeb24c70ea96467e1d705537790dfaa929957))
+
 ## 2.0.8 (2023-04-05)
 
 ## 2.0.7 (2023-03-27)