+++ packages/ui/atomic/create-atomic/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.38.0 (2023-06-13)
+
+### Features
+
+- **organizationendpoints:** add support for organization endpoints for UI commands and projects ([#1289](https://github.com/coveo/cli/issues/1289)) ([3b7653d](https://github.com/coveo/cli/commits/3b7653dbf1b59015afb4575bd265ec0a91b2bcef))
+
+## 1.37.3 (2023-04-21)
+
+### Bug Fixes
+
+- **atomic:** don't need peerDeps ([#1294](https://github.com/coveo/cli/issues/1294)) ([8a7e7a9](https://github.com/coveo/cli/commits/8a7e7a9ae60b3837d2e3820d1af824756e4db549))
+
 ## 1.37.1 (2023-03-27)
 
 ### Bug Fixes